Revert "Ambient breakdown sessions handler (#242)"
This reverts commit 47eccf0bf05e42ae5d29901335fa51de975e9a7d.
