+++ components/backend/handlers/projects.go
@@ -1,5 +1,3 @@
-// Package handlers provides HTTP handlers for the backend API.
-// This file contains CRUD operations for project management.
 package handlers
 
 import (
@@ -7,17 +5,139 @@ import (
 	"fmt"
 	"log"
 	"net/http"
+	"regexp"
+	"strings"
+	"sync"
 	"time"
 
 	"ambient-code-backend/types"
 
 	"github.com/gin-gonic/gin"
+	authv1 "k8s.io/api/authorization/v1"
 	corev1 "k8s.io/api/core/v1"
 	rbacv1 "k8s.io/api/rbac/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
+	"k8s.io/apimachinery/pkg/runtime/schema"
 	k8stypes "k8s.io/apimachinery/pkg/types"
+	"k8s.io/client-go/dynamic"
+	"k8s.io/client-go/kubernetes"
 )
+
+// Package-level variables for project handlers (set from main package)
+var (
+	// GetOpenShiftProjectResource returns the GVR for OpenShift Project resources
+	GetOpenShiftProjectResource func() schema.GroupVersionResource
+	// K8sClientProjects is the backend service account client used for namespace operations
+	// that require elevated permissions (e.g., creating namespaces, assigning roles)
+	K8sClientProjects *kubernetes.Clientset
+	// DynamicClientProjects is the backend SA dynamic client for OpenShift Project operations
+	DynamicClientProjects dynamic.Interface
+)
+
+var (
+	isOpenShiftCache bool
+	isOpenShiftOnce  sync.Once
+)
+
+// Default timeout for Kubernetes API operations
+const defaultK8sTimeout = 10 * time.Second
+
+// Retry configuration constants
+const (
+	projectRetryAttempts     = 5
+	projectRetryInitialDelay = 200 * time.Millisecond
+	projectRetryMaxDelay     = 2 * time.Second
+)
+
+// Kubernetes namespace name validation pattern
+var namespaceNamePattern = regexp.MustCompile(`^[a-z0-9]([-a-z0-9]*[a-z0-9])?$`)
+
+// validateProjectName validates a project/namespace name according to Kubernetes naming rules
+func validateProjectName(name string) error {
+	if name == "" {
+		return fmt.Errorf("project name is required")
+	}
+	if len(name) > 63 {
+		return fmt.Errorf("project name must be 63 characters or less")
+	}
+	if !namespaceNamePattern.MatchString(name) {
+		return fmt.Errorf("project name must be lowercase alphanumeric with hyphens (cannot start or end with hyphen)")
+	}
+	// Reserved namespaces
+	reservedNames := map[string]bool{
+		"default": true, "kube-system": true, "kube-public": true, "kube-node-lease": true,
+		"openshift": true, "openshift-infra": true, "openshift-node": true,
+	}
+	if reservedNames[name] {
+		return fmt.Errorf("project name '%s' is reserved and cannot be used", name)
+	}
+	return nil
+}
+
+// sanitizeForK8sName converts a user subject to a valid Kubernetes resource name
+func sanitizeForK8sName(subject string) string {
+	// Remove system:serviceaccount: prefix if present
+	subject = strings.TrimPrefix(subject, "system:serviceaccount:")
+
+	// Replace invalid characters with hyphens
+	reg := regexp.MustCompile(`[^a-z0-9-]`)
+	sanitized := reg.ReplaceAllString(strings.ToLower(subject), "-")
+
+	// Remove leading/trailing hyphens
+	sanitized = strings.Trim(sanitized, "-")
+
+	// Ensure it doesn't exceed 63 chars (leave room for prefix)
+	if len(sanitized) > 40 {
+		sanitized = sanitized[:40]
+	}
+
+	return sanitized
+}
+
+// isOpenShiftCluster detects if we're running on OpenShift by checking for the project.openshift.io API group
+// Results are cached using sync.Once for thread-safe, race-free initialization
+func isOpenShiftCluster() bool {
+	isOpenShiftOnce.Do(func() {
+		if K8sClientProjects == nil {
+			log.Printf("K8s client not initialized, assuming vanilla Kubernetes")
+			isOpenShiftCache = false
+			return
+		}
+
+		// Try to list API groups and look for project.openshift.io
+		groups, err := K8sClientProjects.Discovery().ServerGroups()
+		if err != nil {
+			log.Printf("Failed to detect OpenShift (assuming vanilla Kubernetes): %v", err)
+			isOpenShiftCache = false
+			return
+		}
+
+		for _, group := range groups.Groups {
+			if group.Name == "project.openshift.io" {
+				log.Printf("Detected OpenShift cluster")
+				isOpenShiftCache = true
+				return
+			}
+		}
+
+		log.Printf("Detected vanilla Kubernetes cluster")
+		isOpenShiftCache = false
+	})
+	return isOpenShiftCache
+}
+
+// GetClusterInfo handles GET /cluster-info
+// Returns information about the cluster type (OpenShift vs vanilla Kubernetes)
+// This endpoint does not require authentication as it's public cluster information
+func GetClusterInfo(c *gin.Context) {
+	isOpenShift := isOpenShiftCluster()
+
+	c.JSON(http.StatusOK, gin.H{
+		"isOpenShift": isOpenShift,
+	})
+}
 
 // ListProjects handles GET /projects
 // Lists Namespaces (both platforms) using backend SA with label selector,
@@ -66,6 +186,35 @@ func ListProjects(c *gin.Context) {
 	}
 
 	c.JSON(http.StatusOK, gin.H{"items": projects})
+}
+
+// projectFromNamespace converts a Kubernetes Namespace to AmbientProject
+// On OpenShift, extracts displayName and description from namespace annotations
+func projectFromNamespace(ns *corev1.Namespace, isOpenShift bool) types.AmbientProject {
+	status := "Active"
+	if ns.Status.Phase != corev1.NamespaceActive {
+		status = string(ns.Status.Phase)
+	}
+
+	displayName := ""
+	description := ""
+
+	// On OpenShift, extract display metadata from annotations
+	if isOpenShift && ns.Annotations != nil {
+		displayName = ns.Annotations["openshift.io/display-name"]
+		description = ns.Annotations["openshift.io/description"]
+	}
+
+	return types.AmbientProject{
+		Name:              ns.Name,
+		DisplayName:       displayName,
+		Description:       description,
+		Labels:            ns.Labels,
+		Annotations:       ns.Annotations,
+		CreationTimestamp: ns.CreationTimestamp.Format(time.RFC3339),
+		Status:            status,
+		IsOpenShift:       isOpenShift,
+	}
 }
 
 // CreateProject handles POST /projects
@@ -233,7 +382,8 @@ func CreateProject(c *gin.Context) {
 			}
 
 			// Update Project annotations with display metadata
-			meta, ok := GetMetadataMap(projObj)
+			var unstruct *unstructured.Unstructured = projObj // Explicit type reference
+			meta, ok := unstruct.Object["metadata"].(map[string]interface{})
 			if !ok || meta == nil {
 				meta = map[string]interface{}{}
 				projObj.Object["metadata"] = meta
@@ -504,31 +654,83 @@ func DeleteProject(c *gin.Context) {
 	c.Status(http.StatusNoContent)
 }
 
-// projectFromNamespace converts a Kubernetes Namespace to AmbientProject
-// On OpenShift, extracts displayName and description from namespace annotations
-func projectFromNamespace(ns *corev1.Namespace, isOpenShift bool) types.AmbientProject {
-	status := "Active"
-	if ns.Status.Phase != corev1.NamespaceActive {
-		status = string(ns.Status.Phase)
+// checkUserCanAccessNamespace uses SelfSubjectAccessReview to verify if user can access a namespace
+// This is the proper Kubernetes-native way - lets RBAC engine determine access from ALL sources
+// (RoleBindings, ClusterRoleBindings, groups, etc.)
+func checkUserCanAccessNamespace(userClient *kubernetes.Clientset, namespace string) (bool, error) {
+	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
+	defer cancel()
+
+	// Check if user can list agenticsessions in the namespace (a good proxy for project access)
+	ssar := &authv1.SelfSubjectAccessReview{
+		Spec: authv1.SelfSubjectAccessReviewSpec{
+			ResourceAttributes: &authv1.ResourceAttributes{
+				Namespace: namespace,
+				Verb:      "list",
+				Group:     "vteam.ambient-code",
+				Resource:  "agenticsessions",
+			},
+		},
 	}
 
-	displayName := ""
-	description := ""
-
-	// On OpenShift, extract display metadata from annotations
-	if isOpenShift && ns.Annotations != nil {
-		displayName = ns.Annotations["openshift.io/display-name"]
-		description = ns.Annotations["openshift.io/description"]
+	result, err := userClient.AuthorizationV1().SelfSubjectAccessReviews().Create(ctx, ssar, v1.CreateOptions{})
+	if err != nil {
+		return false, err
 	}
 
-	return types.AmbientProject{
-		Name:              ns.Name,
-		DisplayName:       displayName,
-		Description:       description,
-		Labels:            ns.Labels,
-		Annotations:       ns.Annotations,
-		CreationTimestamp: ns.CreationTimestamp.Format(time.RFC3339),
-		Status:            status,
-		IsOpenShift:       isOpenShift,
+	return result.Status.Allowed, nil
+}
+
+// getUserSubjectFromContext extracts the user subject from the JWT token in the request
+// Returns subject in format like "user@example.com" or "system:serviceaccount:namespace:name"
+func getUserSubjectFromContext(c *gin.Context) (string, error) {
+	// Try to extract from ServiceAccount first
+	ns, saName, ok := ExtractServiceAccountFromAuth(c)
+	if ok {
+		return fmt.Sprintf("system:serviceaccount:%s:%s", ns, saName), nil
 	}
+
+	// Otherwise try to get from context (set by middleware)
+	if userName, exists := c.Get("userName"); exists && userName != nil {
+		return fmt.Sprintf("%v", userName), nil
+	}
+	if userID, exists := c.Get("userID"); exists && userID != nil {
+		return fmt.Sprintf("%v", userID), nil
+	}
+
+	return "", fmt.Errorf("no user subject found in token")
+}
+
+// getUserSubjectKind returns "ServiceAccount" or "User" based on the subject format
+func getUserSubjectKind(subject string) string {
+	if len(subject) > 22 && subject[:22] == "system:serviceaccount:" {
+		return "ServiceAccount"
+	}
+	return "User"
+}
+
+// getUserSubjectName returns the name part of the subject
+// For ServiceAccount: "system:serviceaccount:namespace:name" -> "name"
+// For User: returns the subject as-is
+func getUserSubjectName(subject string) string {
+	if getUserSubjectKind(subject) == "ServiceAccount" {
+		parts := strings.Split(subject, ":")
+		if len(parts) >= 4 {
+			return parts[3]
+		}
+	}
+	return subject
+}
+
+// getUserSubjectNamespace returns the namespace for ServiceAccount subjects
+// For ServiceAccount: "system:serviceaccount:namespace:name" -> "namespace"
+// For User: returns empty string
+func getUserSubjectNamespace(subject string) string {
+	if getUserSubjectKind(subject) == "ServiceAccount" {
+		parts := strings.Split(subject, ":")
+		if len(parts) >= 3 {
+			return parts[2]
+		}
+	}
+	return ""
 }
