--- components/backend/handlers/rfe_crud.go
@@ -1,436 +1,0 @@
-// Package handlers provides HTTP handlers for the backend API.
-// This file contains CRUD operations for RFE workflows.
-package handlers
-
-import (
-	"fmt"
-	"io"
-	"log"
-	"net/http"
-	"strings"
-	"time"
-
-	"ambient-code-backend/git"
-	"ambient-code-backend/types"
-
-	"github.com/gin-gonic/gin"
-	"k8s.io/apimachinery/pkg/api/errors"
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-// ListProjectRFEWorkflows lists all RFE workflows for a project
-func ListProjectRFEWorkflows(c *gin.Context) {
-	project := c.Param("projectName")
-	var workflows []RFEWorkflow
-	// Prefer CRD list with request-scoped client; fallback to file scan if unavailable or fails
-	gvr := GetRFEWorkflowResource()
-	_, reqDyn := GetK8sClientsForRequest(c)
-	if reqDyn != nil {
-		if list, err := reqDyn.Resource(gvr).Namespace(project).List(c.Request.Context(), v1.ListOptions{LabelSelector: fmt.Sprintf("project=%s", project)}); err == nil {
-			for _, item := range list.Items {
-				wf := RfeFromUnstructured(&item)
-				if wf == nil {
-					continue
-				}
-				workflows = append(workflows, *wf)
-			}
-		}
-	}
-	if workflows == nil {
-		workflows = []RFEWorkflow{}
-	}
-	// Return slim summaries: omit artifacts/agentSessions/phaseResults/status/currentPhase
-	summaries := make([]map[string]interface{}, 0, len(workflows))
-	for _, w := range workflows {
-		item := map[string]interface{}{
-			"id":            w.ID,
-			"title":         w.Title,
-			"description":   w.Description,
-			"branchName":    w.BranchName,
-			"project":       w.Project,
-			"workspacePath": w.WorkspacePath,
-			"createdAt":     w.CreatedAt,
-			"updatedAt":     w.UpdatedAt,
-		}
-		if w.UmbrellaRepo != nil {
-			u := map[string]interface{}{"url": w.UmbrellaRepo.URL}
-			if w.UmbrellaRepo.Branch != nil {
-				u["branch"] = *w.UmbrellaRepo.Branch
-			}
-			item["umbrellaRepo"] = u
-		}
-		if len(w.SupportingRepos) > 0 {
-			repos := make([]map[string]interface{}, 0, len(w.SupportingRepos))
-			for _, r := range w.SupportingRepos {
-				rm := map[string]interface{}{"url": r.URL}
-				if r.Branch != nil {
-					rm["branch"] = *r.Branch
-				}
-				repos = append(repos, rm)
-			}
-			item["supportingRepos"] = repos
-		}
-		summaries = append(summaries, item)
-	}
-	c.JSON(http.StatusOK, gin.H{"workflows": summaries})
-}
-
-// CreateProjectRFEWorkflow creates a new RFE workflow for a project
-func CreateProjectRFEWorkflow(c *gin.Context) {
-	project := c.Param("projectName")
-	var req CreateRFEWorkflowRequest
-	bodyBytes, _ := c.GetRawData()
-	c.Request.Body = io.NopCloser(strings.NewReader(string(bodyBytes)))
-	if err := c.ShouldBindJSON(&req); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Validation failed: " + err.Error()})
-		return
-	}
-	now := time.Now().UTC().Format(time.RFC3339)
-	workflowID := fmt.Sprintf("rfe-%d", time.Now().Unix())
-
-	// Branch name is required and generated by frontend (auto-populated from title, user-editable)
-	// Frontend generates: ambient-{first-three-words-from-title}
-	// Backend only validates that it's not empty and not a protected branch
-	branchName := strings.TrimSpace(req.BranchName)
-	if err := git.ValidateBranchName(branchName); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	// Validate no duplicate repository URLs
-	if err := validateUniqueRepositories(&req.UmbrellaRepo, req.SupportingRepos); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-		return
-	}
-
-	workflow := &RFEWorkflow{
-		ID:              workflowID,
-		Title:           req.Title,
-		Description:     req.Description,
-		BranchName:      branchName,
-		UmbrellaRepo:    &req.UmbrellaRepo,
-		SupportingRepos: req.SupportingRepos,
-		WorkspacePath:   req.WorkspacePath,
-		Project:         project,
-		CreatedAt:       now,
-		UpdatedAt:       now,
-	}
-	_, reqDyn := GetK8sClientsForRequest(c)
-	if err := UpsertProjectRFEWorkflowCR(reqDyn, workflow); err != nil {
-		log.Printf("⚠️ Failed to upsert RFEWorkflow CR: %v", err)
-	}
-
-	// Seeding (spec-kit + agents) is now handled by POST /seed endpoint after creation
-
-	c.JSON(http.StatusCreated, workflow)
-}
-
-// GetProjectRFEWorkflow retrieves a specific RFE workflow by ID
-func GetProjectRFEWorkflow(c *gin.Context) {
-	project := c.Param("projectName")
-	id := c.Param("id")
-	// Try CRD with request-scoped client first
-	gvr := GetRFEWorkflowResource()
-	_, reqDyn := GetK8sClientsForRequest(c)
-	var wf *RFEWorkflow
-	var err error
-	if reqDyn != nil {
-		if item, gerr := reqDyn.Resource(gvr).Namespace(project).Get(c.Request.Context(), id, v1.GetOptions{}); gerr == nil {
-			wf = RfeFromUnstructured(item)
-			err = nil
-		} else {
-			err = gerr
-		}
-	}
-	if err != nil {
-		c.JSON(http.StatusNotFound, gin.H{"error": "Workflow not found"})
-		return
-	}
-	// Return slim object without artifacts/agentSessions/phaseResults/status/currentPhase
-	resp := map[string]interface{}{
-		"id":            wf.ID,
-		"title":         wf.Title,
-		"description":   wf.Description,
-		"branchName":    wf.BranchName,
-		"project":       wf.Project,
-		"workspacePath": wf.WorkspacePath,
-		"createdAt":     wf.CreatedAt,
-		"updatedAt":     wf.UpdatedAt,
-	}
-	if wf.ParentOutcome != nil {
-		resp["parentOutcome"] = *wf.ParentOutcome
-	}
-	if len(wf.JiraLinks) > 0 {
-		links := make([]map[string]interface{}, 0, len(wf.JiraLinks))
-		for _, l := range wf.JiraLinks {
-			links = append(links, map[string]interface{}{"path": l.Path, "jiraKey": l.JiraKey})
-		}
-		resp["jiraLinks"] = links
-	}
-	if wf.UmbrellaRepo != nil {
-		u := map[string]interface{}{"url": wf.UmbrellaRepo.URL}
-		if wf.UmbrellaRepo.Branch != nil {
-			u["branch"] = *wf.UmbrellaRepo.Branch
-		}
-		resp["umbrellaRepo"] = u
-	}
-	if len(wf.SupportingRepos) > 0 {
-		repos := make([]map[string]interface{}, 0, len(wf.SupportingRepos))
-		for _, r := range wf.SupportingRepos {
-			rm := map[string]interface{}{"url": r.URL}
-			if r.Branch != nil {
-				rm["branch"] = *r.Branch
-			}
-			repos = append(repos, rm)
-		}
-		resp["supportingRepos"] = repos
-	}
-	c.JSON(http.StatusOK, resp)
-}
-
-// UpdateProjectRFEWorkflow updates an existing RFE workflow's repository configuration
-// This is primarily used to fix repository URLs before seeding
-func UpdateProjectRFEWorkflow(c *gin.Context) {
-	project := c.Param("projectName")
-	id := c.Param("id")
-
-	var req types.UpdateRFEWorkflowRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Validation failed: " + err.Error()})
-		return
-	}
-
-	// Get the workflow
-	gvr := GetRFEWorkflowResource()
-	_, reqDyn := GetK8sClientsForRequest(c)
-	if reqDyn == nil {
-		c.JSON(http.StatusUnauthorized, gin.H{"error": "Authentication required"})
-		return
-	}
-
-	item, err := reqDyn.Resource(gvr).Namespace(project).Get(c.Request.Context(), id, v1.GetOptions{})
-	if err != nil {
-		if errors.IsNotFound(err) {
-			c.JSON(http.StatusNotFound, gin.H{"error": "Workflow not found"})
-		} else if errors.IsForbidden(err) {
-			c.JSON(http.StatusForbidden, gin.H{"error": "You don't have permission to access this workflow"})
-		} else {
-			log.Printf("Failed to get workflow %s: %v", id, err)
-			c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to retrieve workflow"})
-		}
-		return
-	}
-
-	wf := RfeFromUnstructured(item)
-	if wf == nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid workflow"})
-		return
-	}
-
-	// Validate no duplicate repository URLs if repositories are being updated
-	if req.UmbrellaRepo != nil || req.SupportingRepos != nil {
-		umbrellaRepo := req.UmbrellaRepo
-		if umbrellaRepo == nil {
-			umbrellaRepo = wf.UmbrellaRepo
-		}
-		supportingRepos := req.SupportingRepos
-		if supportingRepos == nil {
-			supportingRepos = wf.SupportingRepos
-		}
-		if err := validateUniqueRepositories(umbrellaRepo, supportingRepos); err != nil {
-			c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
-			return
-		}
-	}
-
-	// Update the CR
-	obj := item.DeepCopy()
-	spec, ok := GetSpecMap(obj)
-	if !ok {
-		spec = make(map[string]interface{})
-		obj.Object["spec"] = spec
-	}
-
-	// Update fields if provided
-	if req.Title != nil {
-		spec["title"] = *req.Title
-	}
-	if req.Description != nil {
-		spec["description"] = *req.Description
-	}
-	if req.UmbrellaRepo != nil {
-		spec["umbrellaRepo"] = map[string]interface{}{
-			"url":    req.UmbrellaRepo.URL,
-			"branch": req.UmbrellaRepo.Branch,
-		}
-	}
-	if req.SupportingRepos != nil {
-		repos := make([]interface{}, len(req.SupportingRepos))
-		for i, r := range req.SupportingRepos {
-			repos[i] = map[string]interface{}{
-				"url":    r.URL,
-				"branch": r.Branch,
-			}
-		}
-		spec["supportingRepos"] = repos
-	}
-	if req.ParentOutcome != nil {
-		spec["parentOutcome"] = *req.ParentOutcome
-	}
-
-	// Update the CR
-	updated, err := reqDyn.Resource(gvr).Namespace(project).Update(c.Request.Context(), obj, v1.UpdateOptions{})
-	if err != nil {
-		log.Printf("Failed to update RFEWorkflow CR: %v", err)
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to update workflow"})
-		return
-	}
-
-	// Convert back to RFEWorkflow type
-	updatedWf := RfeFromUnstructured(updated)
-	if updatedWf == nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to parse updated workflow"})
-		return
-	}
-
-	c.JSON(http.StatusOK, updatedWf)
-}
-
-// GetProjectRFEWorkflowSummary computes derived phase/status and progress based on workspace files and linked sessions
-// GET /api/projects/:projectName/rfe-workflows/:id/summary
-func GetProjectRFEWorkflowSummary(c *gin.Context) {
-	project := c.Param("projectName")
-	id := c.Param("id")
-
-	// Determine workspace and expected files
-	// workspace content removed
-	specsItems := []contentListItem{}
-
-	hasSpec := false
-	hasPlan := false
-	hasTasks := false
-
-	// helper to scan a list for target filenames
-	scanFor := func(items []contentListItem) (bool, bool, bool) {
-		s, p, t := false, false, false
-		for _, it := range items {
-			if it.IsDir {
-				continue
-			}
-			switch strings.ToLower(it.Name) {
-			case "spec.md":
-				s = true
-			case "plan.md":
-				p = true
-			case "tasks.md":
-				t = true
-			}
-		}
-		return s, p, t
-	}
-
-	// First check directly under specs/
-	if len(specsItems) > 0 {
-		s, p, t := scanFor(specsItems)
-		hasSpec, hasPlan, hasTasks = s, p, t
-		// If not found, check first subfolder under specs/
-		if !hasSpec && !hasPlan && !hasTasks {
-			for _, it := range specsItems {
-				if it.IsDir {
-					subItems := []contentListItem{}
-					s2, p2, t2 := scanFor(subItems)
-					hasSpec, hasPlan, hasTasks = s2, p2, t2
-					break
-				}
-			}
-		}
-	}
-
-	// Sessions: find linked sessions and compute running/failed flags
-	gvr := GetAgenticSessionV1Alpha1Resource()
-	_, reqDyn := GetK8sClientsForRequest(c)
-	anyRunning := false
-	anyFailed := false
-	if reqDyn != nil {
-		selector := fmt.Sprintf("rfe-workflow=%s,project=%s", id, project)
-		if list, err := reqDyn.Resource(gvr).Namespace(project).List(c.Request.Context(), v1.ListOptions{LabelSelector: selector}); err == nil {
-			for _, item := range list.Items {
-				status, ok := GetStatusMap(&item)
-				if !ok {
-					status = make(map[string]interface{})
-				}
-				phaseStr := strings.ToLower(fmt.Sprintf("%v", status["phase"]))
-				if phaseStr == "running" || phaseStr == "creating" || phaseStr == "pending" {
-					anyRunning = true
-				}
-				if phaseStr == "failed" || phaseStr == "error" {
-					anyFailed = true
-				}
-			}
-		}
-	}
-
-	// Derive phase and status
-	var phase string
-	switch {
-	case !hasSpec && !hasPlan && !hasTasks:
-		phase = "pre"
-	case !hasSpec:
-		phase = "specify"
-	case !hasPlan:
-		phase = "plan"
-	case !hasTasks:
-		phase = "tasks"
-	default:
-		phase = "completed"
-	}
-
-	status := "not started"
-	if anyRunning {
-		status = "running"
-	} else if hasSpec || hasPlan || hasTasks {
-		status = "in progress"
-	}
-	if hasSpec && hasPlan && hasTasks && !anyRunning {
-		status = "completed"
-	}
-	if anyFailed && status != "running" {
-		status = "attention"
-	}
-
-	progress := float64(0)
-	done := 0
-	if hasSpec {
-		done++
-	}
-	if hasPlan {
-		done++
-	}
-	if hasTasks {
-		done++
-	}
-	progress = float64(done) / 3.0 * 100.0
-
-	c.JSON(http.StatusOK, gin.H{
-		"phase":    phase,
-		"status":   status,
-		"progress": progress,
-		"files": gin.H{
-			"spec":  hasSpec,
-			"plan":  hasPlan,
-			"tasks": hasTasks,
-		},
-	})
-}
-
-// DeleteProjectRFEWorkflow deletes an RFE workflow
-func DeleteProjectRFEWorkflow(c *gin.Context) {
-	id := c.Param("id")
-	// Delete CR
-	gvr := GetRFEWorkflowResource()
-	_, reqDyn := GetK8sClientsForRequest(c)
-	if reqDyn != nil {
-		_ = reqDyn.Resource(gvr).Namespace(c.Param("projectName")).Delete(c.Request.Context(), id, v1.DeleteOptions{})
-	}
-	c.JSON(http.StatusOK, gin.H{"message": "Workflow deleted successfully"})
-}
