--- components/backend/handlers/rfe_sessions.go
@@ -1,126 +1,0 @@
-// Package handlers provides HTTP handlers for the backend API.
-// This file contains session management operations for RFE workflows.
-package handlers
-
-import (
-	"fmt"
-	"net/http"
-
-	"github.com/gin-gonic/gin"
-	"k8s.io/apimachinery/pkg/api/errors"
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-// ListProjectRFEWorkflowSessions lists sessions linked to a project-scoped RFE workflow by label selector
-func ListProjectRFEWorkflowSessions(c *gin.Context) {
-	project := c.Param("projectName")
-	id := c.Param("id")
-	gvr := GetAgenticSessionV1Alpha1Resource()
-	selector := fmt.Sprintf("rfe-workflow=%s,project=%s", id, project)
-	_, reqDyn := GetK8sClientsForRequest(c)
-	if reqDyn == nil {
-		c.JSON(http.StatusUnauthorized, gin.H{"error": "Missing or invalid user token"})
-		return
-	}
-	list, err := reqDyn.Resource(gvr).Namespace(project).List(c.Request.Context(), v1.ListOptions{LabelSelector: selector})
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to list sessions", "details": err.Error()})
-		return
-	}
-
-	// Return full session objects for UI
-	sessions := make([]map[string]interface{}, 0, len(list.Items))
-	for _, item := range list.Items {
-		sessions = append(sessions, item.Object)
-	}
-	c.JSON(http.StatusOK, gin.H{"sessions": sessions})
-}
-
-// AddProjectRFEWorkflowSession adds/links an existing session to an RFE by applying labels
-func AddProjectRFEWorkflowSession(c *gin.Context) {
-	project := c.Param("projectName")
-	id := c.Param("id")
-	var req rfeLinkSessionRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid request: " + err.Error()})
-		return
-	}
-	if req.ExistingName == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "existingName is required for linking in this version"})
-		return
-	}
-	gvr := GetAgenticSessionV1Alpha1Resource()
-	_, reqDyn := GetK8sClientsForRequest(c)
-	if reqDyn == nil {
-		c.JSON(http.StatusUnauthorized, gin.H{"error": "Missing or invalid user token"})
-		return
-	}
-	obj, err := reqDyn.Resource(gvr).Namespace(project).Get(c.Request.Context(), req.ExistingName, v1.GetOptions{})
-	if err != nil {
-		if errors.IsNotFound(err) {
-			c.JSON(http.StatusNotFound, gin.H{"error": "Session not found"})
-			return
-		}
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to fetch session", "details": err.Error()})
-		return
-	}
-	meta, ok := GetMetadataMap(obj)
-	if !ok {
-		meta = make(map[string]interface{})
-	}
-	labels, _ := meta["labels"].(map[string]interface{})
-	if labels == nil {
-		labels = map[string]interface{}{}
-		meta["labels"] = labels
-	}
-	labels["project"] = project
-	labels["rfe-workflow"] = id
-	if req.Phase != "" {
-		labels["rfe-phase"] = req.Phase
-	}
-	// Update the resource
-	updated, err := reqDyn.Resource(gvr).Namespace(project).Update(c.Request.Context(), obj, v1.UpdateOptions{})
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to update session labels", "details": err.Error()})
-		return
-	}
-	_ = updated
-	c.JSON(http.StatusOK, gin.H{"message": "Session linked to RFE", "session": req.ExistingName})
-}
-
-// RemoveProjectRFEWorkflowSession removes/unlinks a session from an RFE by clearing linkage labels (non-destructive)
-func RemoveProjectRFEWorkflowSession(c *gin.Context) {
-	project := c.Param("projectName")
-	_ = project // currently unused but kept for parity/logging if needed
-	id := c.Param("id")
-	sessionName := c.Param("sessionName")
-	gvr := GetAgenticSessionV1Alpha1Resource()
-	_, reqDyn := GetK8sClientsForRequest(c)
-	if reqDyn == nil {
-		c.JSON(http.StatusUnauthorized, gin.H{"error": "Missing or invalid user token"})
-		return
-	}
-	obj, err := reqDyn.Resource(gvr).Namespace(project).Get(c.Request.Context(), sessionName, v1.GetOptions{})
-	if err != nil {
-		if errors.IsNotFound(err) {
-			c.JSON(http.StatusNotFound, gin.H{"error": "Session not found"})
-			return
-		}
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to fetch session", "details": err.Error()})
-		return
-	}
-	meta, ok := GetMetadataMap(obj)
-	if !ok {
-		meta = make(map[string]interface{})
-	}
-	labels, _ := meta["labels"].(map[string]interface{})
-	if labels != nil {
-		delete(labels, "rfe-workflow")
-		delete(labels, "rfe-phase")
-	}
-	if _, err := reqDyn.Resource(gvr).Namespace(project).Update(c.Request.Context(), obj, v1.UpdateOptions{}); err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to update session labels", "details": err.Error()})
-		return
-	}
-	c.JSON(http.StatusOK, gin.H{"message": "Session unlinked from RFE", "session": sessionName, "rfe": id})
-}
