--- components/backend/handlers/sessions_git.go
@@ -1,475 +1,0 @@
-// Package handlers provides HTTP handlers for the backend API.
-// This file contains git/GitHub operations for agentic sessions.
-package handlers
-
-import (
-	"context"
-	"encoding/json"
-	"fmt"
-	"io"
-	"log"
-	"net/http"
-	"net/url"
-	"strings"
-	"time"
-
-	"github.com/gin-gonic/gin"
-	authnv1 "k8s.io/api/authentication/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/client-go/dynamic"
-)
-
-// MintSessionGitHubToken generates a GitHub token for authenticated session operations.
-// POST /api/projects/:projectName/agentic-sessions/:sessionName/github/token
-func MintSessionGitHubToken(c *gin.Context) {
-	project := c.Param("projectName")
-	sessionName := c.Param("sessionName")
-
-	rawAuth := strings.TrimSpace(c.GetHeader("Authorization"))
-	if rawAuth == "" {
-		c.JSON(http.StatusUnauthorized, gin.H{"error": "missing Authorization header"})
-		return
-	}
-	parts := strings.SplitN(rawAuth, " ", 2)
-	if len(parts) != 2 || !strings.EqualFold(parts[0], "Bearer") {
-		c.JSON(http.StatusUnauthorized, gin.H{"error": "invalid Authorization header"})
-		return
-	}
-	token := strings.TrimSpace(parts[1])
-	if token == "" {
-		c.JSON(http.StatusUnauthorized, gin.H{"error": "empty token"})
-		return
-	}
-
-	// TokenReview using default audience (works with standard SA tokens)
-	tr := &authnv1.TokenReview{Spec: authnv1.TokenReviewSpec{Token: token}}
-	rv, err := K8sClient.AuthenticationV1().TokenReviews().Create(c.Request.Context(), tr, v1.CreateOptions{})
-	if err != nil {
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "token review failed"})
-		return
-	}
-	if rv.Status.Error != "" || !rv.Status.Authenticated {
-		c.JSON(http.StatusUnauthorized, gin.H{"error": "unauthenticated"})
-		return
-	}
-	subj := strings.TrimSpace(rv.Status.User.Username)
-	const pfx = "system:serviceaccount:"
-	if !strings.HasPrefix(subj, pfx) {
-		c.JSON(http.StatusForbidden, gin.H{"error": "subject is not a service account"})
-		return
-	}
-	rest := strings.TrimPrefix(subj, pfx)
-	segs := strings.SplitN(rest, ":", 2)
-	if len(segs) != 2 {
-		c.JSON(http.StatusForbidden, gin.H{"error": "invalid service account subject"})
-		return
-	}
-	nsFromToken, saFromToken := segs[0], segs[1]
-	if nsFromToken != project {
-		c.JSON(http.StatusForbidden, gin.H{"error": "namespace mismatch"})
-		return
-	}
-
-	// Load session and verify SA matches annotation
-	gvr := GetAgenticSessionV1Alpha1Resource()
-	obj, err := DynamicClient.Resource(gvr).Namespace(project).Get(c.Request.Context(), sessionName, v1.GetOptions{})
-	if err != nil {
-		if errors.IsNotFound(err) {
-			c.JSON(http.StatusNotFound, gin.H{"error": "session not found"})
-			return
-		}
-		c.JSON(http.StatusInternalServerError, gin.H{"error": "failed to read session"})
-		return
-	}
-	meta, ok := GetMetadataMap(obj)
-	if !ok {
-		meta = make(map[string]interface{})
-	}
-	anns, _ := meta["annotations"].(map[string]interface{})
-	expectedSA := ""
-	if anns != nil {
-		if v, ok := anns["ambient-code.io/runner-sa"].(string); ok {
-			expectedSA = strings.TrimSpace(v)
-		}
-	}
-	if expectedSA == "" || expectedSA != saFromToken {
-		c.JSON(http.StatusForbidden, gin.H{"error": "service account not authorized for session"})
-		return
-	}
-
-	// Read authoritative userId from spec.userContext.userId
-	spec, ok := GetSpecMap(obj)
-	if !ok {
-		spec = make(map[string]interface{})
-	}
-	userId := ""
-	if spec != nil {
-		if uc, ok := spec["userContext"].(map[string]interface{}); ok {
-			if v, ok := uc["userId"].(string); ok {
-				userId = strings.TrimSpace(v)
-			}
-		}
-	}
-	if userId == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "session missing user context"})
-		return
-	}
-
-	// Get GitHub token (GitHub App or PAT fallback via project runner secret)
-	tokenStr, err := GetGitHubToken(c.Request.Context(), K8sClient, DynamicClient, project, userId)
-	if err != nil {
-		c.JSON(http.StatusBadGateway, gin.H{"error": err.Error()})
-		return
-	}
-	// Note: PATs don't have expiration, so we omit expiresAt for simplicity
-	// Runners should treat all tokens as short-lived and request new ones as needed
-	c.JSON(http.StatusOK, gin.H{"token": tokenStr})
-}
-
-// setRepoStatus updates the status of a repository in the session CR.
-func setRepoStatus(dyn dynamic.Interface, project, sessionName string, repoIndex int, newStatus string) error {
-	gvr := GetAgenticSessionV1Alpha1Resource()
-	item, err := dyn.Resource(gvr).Namespace(project).Get(context.TODO(), sessionName, v1.GetOptions{})
-	if err != nil {
-		return err
-	}
-
-	// Get repo name from spec.repos[repoIndex]
-	spec, ok := GetSpecMap(item)
-	if !ok {
-		spec = make(map[string]interface{})
-	}
-	specRepos, _ := spec["repos"].([]interface{})
-	if repoIndex < 0 || repoIndex >= len(specRepos) {
-		return fmt.Errorf("repo index out of range")
-	}
-	specRepo, _ := specRepos[repoIndex].(map[string]interface{})
-	repoName := ""
-	if name, ok := specRepo["name"].(string); ok {
-		repoName = name
-	} else if input, ok := specRepo["input"].(map[string]interface{}); ok {
-		if url, ok := input["url"].(string); ok {
-			repoName = DeriveRepoFolderFromURL(url)
-		}
-	}
-	if repoName == "" {
-		repoName = fmt.Sprintf("repo-%d", repoIndex)
-	}
-
-	// Ensure status.repos exists
-	if item.Object["status"] == nil {
-		item.Object["status"] = make(map[string]interface{})
-	}
-	status, ok := GetStatusMap(item)
-	if !ok {
-		status = make(map[string]interface{})
-		item.Object["status"] = status
-	}
-	statusRepos, _ := status["repos"].([]interface{})
-	if statusRepos == nil {
-		statusRepos = []interface{}{}
-	}
-
-	// Find or create status entry for this repo
-	repoStatus := map[string]interface{}{
-		"name":         repoName,
-		"status":       newStatus,
-		"last_updated": time.Now().Format(time.RFC3339),
-	}
-
-	// Update existing or append new
-	found := false
-	for i, r := range statusRepos {
-		if rm, ok := r.(map[string]interface{}); ok {
-			if n, ok := rm["name"].(string); ok && n == repoName {
-				rm["status"] = newStatus
-				rm["last_updated"] = time.Now().Format(time.RFC3339)
-				statusRepos[i] = rm
-				found = true
-				break
-			}
-		}
-	}
-	if !found {
-		statusRepos = append(statusRepos, repoStatus)
-	}
-
-	status["repos"] = statusRepos
-	item.Object["status"] = status
-
-	updated, err := dyn.Resource(gvr).Namespace(project).UpdateStatus(context.TODO(), item, v1.UpdateOptions{})
-	if err != nil {
-		log.Printf("setRepoStatus: update failed project=%s session=%s repoIndex=%d status=%s err=%v", project, sessionName, repoIndex, newStatus, err)
-		return err
-	}
-	if updated != nil {
-		log.Printf("setRepoStatus: update ok project=%s session=%s repo=%s status=%s", project, sessionName, repoName, newStatus)
-	}
-	return nil
-}
-
-// PushSessionRepo pushes changes from a session workspace to the output repository.
-// POST /api/projects/:projectName/agentic-sessions/:sessionName/github/push
-func PushSessionRepo(c *gin.Context) {
-	project := c.Param("projectName")
-	session := c.Param("sessionName")
-
-	var body struct {
-		RepoIndex     int    `json:"repoIndex"`
-		CommitMessage string `json:"commitMessage"`
-	}
-	if err := c.BindJSON(&body); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid JSON body"})
-		return
-	}
-	log.Printf("pushSessionRepo: request project=%s session=%s repoIndex=%d commitLen=%d", project, session, body.RepoIndex, len(strings.TrimSpace(body.CommitMessage)))
-
-	// Resolve the correct content service (temp-content for completed, ambient-content for running)
-	serviceName := ResolveContentServiceName(c, project, session)
-	endpoint := fmt.Sprintf("http://%s.%s.svc:8080", serviceName, project)
-	log.Printf("pushSessionRepo: using service %s", serviceName)
-
-	// Simplified: 1) get session; 2) compute repoPath from INPUT repo folder; 3) get output url/branch; 4) proxy
-	resolvedRepoPath := ""
-	// default branch when not defined on output
-	resolvedBranch := fmt.Sprintf("sessions/%s", session)
-	resolvedOutputURL := ""
-	if _, reqDyn := GetK8sClientsForRequest(c); reqDyn != nil {
-		gvr := GetAgenticSessionV1Alpha1Resource()
-		obj, err := reqDyn.Resource(gvr).Namespace(project).Get(c.Request.Context(), session, v1.GetOptions{})
-		if err != nil {
-			c.JSON(http.StatusBadRequest, gin.H{"error": "failed to read session"})
-			return
-		}
-		spec, ok := GetSpecMap(obj)
-		if !ok {
-			spec = make(map[string]interface{})
-		}
-		repos, _ := spec["repos"].([]interface{})
-		if body.RepoIndex < 0 || body.RepoIndex >= len(repos) {
-			c.JSON(http.StatusBadRequest, gin.H{"error": "invalid repo index"})
-			return
-		}
-		rm, _ := repos[body.RepoIndex].(map[string]interface{})
-		// Derive repoPath from input URL folder name
-		if in, ok := rm["input"].(map[string]interface{}); ok {
-			if urlv, ok2 := in["url"].(string); ok2 && strings.TrimSpace(urlv) != "" {
-				folder := DeriveRepoFolderFromURL(strings.TrimSpace(urlv))
-				if folder != "" {
-					resolvedRepoPath = fmt.Sprintf("/sessions/%s/workspace/%s", session, folder)
-				}
-			}
-		}
-		if out, ok := rm["output"].(map[string]interface{}); ok {
-			if urlv, ok2 := out["url"].(string); ok2 && strings.TrimSpace(urlv) != "" {
-				resolvedOutputURL = strings.TrimSpace(urlv)
-			}
-			if bs, ok2 := out["branch"].(string); ok2 && strings.TrimSpace(bs) != "" {
-				resolvedBranch = strings.TrimSpace(bs)
-			} else if bv, ok2 := out["branch"].(*string); ok2 && bv != nil && strings.TrimSpace(*bv) != "" {
-				resolvedBranch = strings.TrimSpace(*bv)
-			}
-		}
-	} else {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "no dynamic client"})
-		return
-	}
-	// If input URL missing or unparsable, fall back to numeric index path (last resort)
-	if strings.TrimSpace(resolvedRepoPath) == "" {
-		if body.RepoIndex >= 0 {
-			resolvedRepoPath = fmt.Sprintf("/sessions/%s/workspace/%d", session, body.RepoIndex)
-		} else {
-			resolvedRepoPath = fmt.Sprintf("/sessions/%s/workspace", session)
-		}
-	}
-	if strings.TrimSpace(resolvedOutputURL) == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "missing output repo url"})
-		return
-	}
-	log.Printf("pushSessionRepo: resolved repoPath=%q outputUrl=%q branch=%q", resolvedRepoPath, resolvedOutputURL, resolvedBranch)
-
-	payload := map[string]interface{}{
-		"repoPath":      resolvedRepoPath,
-		"commitMessage": body.CommitMessage,
-		"branch":        resolvedBranch,
-		"outputRepoUrl": resolvedOutputURL,
-	}
-	b, _ := json.Marshal(payload)
-	req, _ := http.NewRequestWithContext(c.Request.Context(), http.MethodPost, endpoint+"/content/github/push", strings.NewReader(string(b)))
-	if v := c.GetHeader("Authorization"); v != "" {
-		req.Header.Set("Authorization", v)
-	}
-	if v := c.GetHeader("X-Forwarded-Access-Token"); v != "" {
-		req.Header.Set("X-Forwarded-Access-Token", v)
-	}
-	req.Header.Set("Content-Type", "application/json")
-
-	// Attach short-lived GitHub token for one-shot authenticated push
-	if reqK8s, reqDyn := GetK8sClientsForRequest(c); reqK8s != nil {
-		// Load session to get authoritative userId
-		gvr := GetAgenticSessionV1Alpha1Resource()
-		obj, err := reqDyn.Resource(gvr).Namespace(project).Get(c.Request.Context(), session, v1.GetOptions{})
-		if err == nil {
-			spec, ok := GetSpecMap(obj)
-			if !ok {
-				spec = make(map[string]interface{})
-			}
-			userId := ""
-			if spec != nil {
-				if uc, ok := spec["userContext"].(map[string]interface{}); ok {
-					if v, ok := uc["userId"].(string); ok {
-						userId = strings.TrimSpace(v)
-					}
-				}
-			}
-			if userId != "" {
-				if tokenStr, err := GetGitHubToken(c.Request.Context(), reqK8s, reqDyn, project, userId); err == nil && strings.TrimSpace(tokenStr) != "" {
-					req.Header.Set("X-GitHub-Token", tokenStr)
-					log.Printf("pushSessionRepo: attached short-lived GitHub token for project=%s session=%s", project, session)
-				} else if err != nil {
-					log.Printf("pushSessionRepo: failed to resolve GitHub token: %v", err)
-				}
-			} else {
-				log.Printf("pushSessionRepo: session %s/%s missing userContext.userId; proceeding without token", project, session)
-			}
-		} else {
-			log.Printf("pushSessionRepo: failed to read session for token attach: %v", err)
-		}
-	}
-
-	log.Printf("pushSessionRepo: proxy push project=%s session=%s repoIndex=%d repoPath=%s endpoint=%s", project, session, body.RepoIndex, resolvedRepoPath, endpoint+"/content/github/push")
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		c.JSON(http.StatusBadGateway, gin.H{"error": err.Error()})
-		return
-	}
-	defer resp.Body.Close()
-	bodyBytes, _ := io.ReadAll(resp.Body)
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		log.Printf("pushSessionRepo: content returned status=%d body.snip=%q", resp.StatusCode, func() string {
-			s := string(bodyBytes)
-			if len(s) > 1500 {
-				return s[:1500] + "..."
-			}
-			return s
-		}())
-		c.Data(resp.StatusCode, "application/json", bodyBytes)
-		return
-	}
-	if _, reqDyn := GetK8sClientsForRequest(c); reqDyn != nil {
-		log.Printf("pushSessionRepo: setting repo status to 'pushed' for repoIndex=%d", body.RepoIndex)
-		if err := setRepoStatus(reqDyn, project, session, body.RepoIndex, "pushed"); err != nil {
-			log.Printf("pushSessionRepo: setRepoStatus failed project=%s session=%s repoIndex=%d err=%v", project, session, body.RepoIndex, err)
-		}
-	} else {
-		log.Printf("pushSessionRepo: no dynamic client; cannot set repo status project=%s session=%s", project, session)
-	}
-	log.Printf("pushSessionRepo: content push succeeded status=%d body.len=%d", resp.StatusCode, len(bodyBytes))
-	c.Data(http.StatusOK, "application/json", bodyBytes)
-}
-
-// AbandonSessionRepo marks a repository as abandoned without pushing changes.
-// POST /api/projects/:projectName/agentic-sessions/:sessionName/github/abandon
-func AbandonSessionRepo(c *gin.Context) {
-	project := c.Param("projectName")
-	session := c.Param("sessionName")
-	var body struct {
-		RepoIndex int    `json:"repoIndex"`
-		RepoPath  string `json:"repoPath"`
-	}
-	if err := c.BindJSON(&body); err != nil {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "invalid JSON body"})
-		return
-	}
-
-	// Resolve the correct content service (temp-content for completed, ambient-content for running)
-	serviceName := ResolveContentServiceName(c, project, session)
-	endpoint := fmt.Sprintf("http://%s.%s.svc:8080", serviceName, project)
-	log.Printf("AbandonSessionRepo: using service %s", serviceName)
-	repoPath := strings.TrimSpace(body.RepoPath)
-	if repoPath == "" {
-		if body.RepoIndex >= 0 {
-			repoPath = fmt.Sprintf("/sessions/%s/workspace/%d", session, body.RepoIndex)
-		} else {
-			repoPath = fmt.Sprintf("/sessions/%s/workspace", session)
-		}
-	}
-	payload := map[string]interface{}{
-		"repoPath": repoPath,
-	}
-	b, _ := json.Marshal(payload)
-	req, _ := http.NewRequestWithContext(c.Request.Context(), http.MethodPost, endpoint+"/content/github/abandon", strings.NewReader(string(b)))
-	if v := c.GetHeader("Authorization"); v != "" {
-		req.Header.Set("Authorization", v)
-	}
-	if v := c.GetHeader("X-Forwarded-Access-Token"); v != "" {
-		req.Header.Set("X-Forwarded-Access-Token", v)
-	}
-	req.Header.Set("Content-Type", "application/json")
-	log.Printf("abandonSessionRepo: proxy abandon project=%s session=%s repoIndex=%d repoPath=%s", project, session, body.RepoIndex, repoPath)
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		c.JSON(http.StatusBadGateway, gin.H{"error": err.Error()})
-		return
-	}
-	defer resp.Body.Close()
-	bodyBytes, _ := io.ReadAll(resp.Body)
-	if resp.StatusCode < 200 || resp.StatusCode >= 300 {
-		log.Printf("abandonSessionRepo: content returned status=%d body=%s", resp.StatusCode, string(bodyBytes))
-		c.Data(resp.StatusCode, "application/json", bodyBytes)
-		return
-	}
-	if _, reqDyn := GetK8sClientsForRequest(c); reqDyn != nil {
-		if err := setRepoStatus(reqDyn, project, session, body.RepoIndex, "abandoned"); err != nil {
-			log.Printf("abandonSessionRepo: setRepoStatus failed project=%s session=%s repoIndex=%d err=%v", project, session, body.RepoIndex, err)
-		}
-	} else {
-		log.Printf("abandonSessionRepo: no dynamic client; cannot set repo status project=%s session=%s", project, session)
-	}
-	c.Data(http.StatusOK, "application/json", bodyBytes)
-}
-
-// DiffSessionRepo returns git diff for changes in a session workspace repository.
-// GET /api/projects/:projectName/agentic-sessions/:sessionName/github/diff
-func DiffSessionRepo(c *gin.Context) {
-	project := c.Param("projectName")
-	session := c.Param("sessionName")
-	repoIndexStr := strings.TrimSpace(c.Query("repoIndex"))
-	repoPath := strings.TrimSpace(c.Query("repoPath"))
-	if repoPath == "" && repoIndexStr != "" {
-		repoPath = fmt.Sprintf("/sessions/%s/workspace/%s", session, repoIndexStr)
-	}
-	if repoPath == "" {
-		c.JSON(http.StatusBadRequest, gin.H{"error": "missing repoPath/repoIndex"})
-		return
-	}
-
-	// Resolve the correct content service (temp-content for completed, ambient-content for running)
-	serviceName := ResolveContentServiceName(c, project, session)
-	endpoint := fmt.Sprintf("http://%s.%s.svc:8080", serviceName, project)
-	log.Printf("DiffSessionRepo: using service %s", serviceName)
-	urlStr := fmt.Sprintf("%s/content/github/diff?repoPath=%s", endpoint, url.QueryEscape(repoPath))
-	req, _ := http.NewRequestWithContext(c.Request.Context(), http.MethodGet, urlStr, nil)
-	if v := c.GetHeader("Authorization"); v != "" {
-		req.Header.Set("Authorization", v)
-	}
-	if v := c.GetHeader("X-Forwarded-Access-Token"); v != "" {
-		req.Header.Set("X-Forwarded-Access-Token", v)
-	}
-	resp, err := http.DefaultClient.Do(req)
-	if err != nil {
-		c.JSON(http.StatusOK, gin.H{
-			"files": gin.H{
-				"added":   0,
-				"removed": 0,
-			},
-			"total_added":   0,
-			"total_removed": 0,
-		})
-		return
-	}
-	defer resp.Body.Close()
-	bodyBytes, _ := io.ReadAll(resp.Body)
-	c.Data(resp.StatusCode, resp.Header.Get("Content-Type"), bodyBytes)
-}
